# frontend/src/lib/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing blips, technologies, users, and user technologies
  title: Tech Radar API
  contact: {}
  version: "1.0"
paths:
...
  /user-technologies:
    post:
      description: Assign a technology to a user with a ring
      tags:
        - user-technologies
      summary: Create a user technology
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handlers.CreateUserTechnologyRequest"
        description: UserTechnology data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.UserTechnology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
  "/user-technologies/user/{user_id}":
    get:
      description: Get all technologies assigned to a user
      tags:
        - user-technologies
      summary: Get technologies for a user
      parameters:
        - description: User ID
          name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/handlers.UserTechnology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
  "/user-technologies/{id}":
    get:
      description: Get user technology by ID
      tags:
        - user-technologies
      summary: Get a user technology
      parameters:
        - description: UserTechnology ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.UserTechnology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
    put:
      description: Update user technology ring
      tags:
        - user-technologies
      summary: Update a user technology
      parameters:
        - description: UserTechnology ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handlers.UpdateUserTechnologyRequest"
        description: UserTechnology data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.UserTechnology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
    delete:
      description: Delete user technology by ID
      tags:
        - user-technologies
      summary: Delete a user technology
      parameters:
        - description: UserTechnology ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
...
servers:
  - url: //localhost:8080
components:
  schemas:
...
    handlers.CreateUserTechnologyRequest:
      type: object
      properties:
        id:
          type: string
        ring_id:
          type: integer
        technology_id:
          type: string
        user_id:
          type: string
    handlers.Error:
      type: object
      properties:
        message:
          type: string
...
    handlers.UpdateUserTechnologyRequest:
      type: object
      properties:
        ring_id:
          type: integer
        technology_id:
          type: string
        user_id:
          type: string
...
    handlers.UserTechnology:
      type: object
      properties:
        created_at:
          type: string
          example: 2026-04-05T12:00:00Z
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        ring_id:
          type: integer
          example: 2
        technology_id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        updated_at:
          type: string
          example: 2026-04-05T12:00:00Z
        user_id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000